+++ decode/emitter/emitter.py
@@ -97,6 +97,9 @@ class EmitterSet:
             xy_unit: Unit of the x and y coordinate.
             px_size: Pixel size for unit conversion. If not specified,
                 derived attributes (xyz_px and xyz_nm) can not be accessed
+            iframe: select by frame_index (e.g. emitter.iframe[0:10] will output
+                emitters on frames 0-9
+            icode: select by code (similiar to iframe)
     """
 
     _data_holder = {
@@ -182,6 +185,7 @@ class EmitterSet:
         self.xy_unit = xy_unit
         self.px_size = px_size
         self.iframe = slicing.SliceForward(self._iframe_hook)
+        self.icode = slicing.SliceForward(self._icode_hook)
         self._sorted = False
 
         if self.px_size is not None:
@@ -750,6 +754,19 @@ class EmitterSet:
             return self.get_subset_frame(item, item + 1)
 
         raise NotImplementedError
+
+    def _icode_hook(self, item):
+        if isinstance(item, slice):
+            if item.step is not None:
+                raise NotImplementedError("Step argument in slicing not supported.")
+            item = (item.start, item.stop)
+        elif isinstance(item, int):
+            item = (item, item + 1)
+        else:
+            raise NotImplementedError
+
+        ix = (self.code >= item[0]) * (self.code < item[1])
+        return self[ix]
 
     def chunks(self, chunks: int) -> list:
         """